ui: Proxmox Backup: allow to autogenerate an encryption key
a bit more complicated than wished, but if we want to have at least a
bsaic OK user experience we need to give good feedback that the key
will be deleted.

We definitively want to extend this, allow uploading their own,
export if root@pam, only move not delete on "removal", for safety.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>