ui: ceph status: use two column layout for 1600+ window width
Looks already OK at that size, and one gets a better overview.

We have a slightly complex layout here (to columns which should be
above each other) so we cannot just use the generic helper, but
that's OK here - it *is* a special view.

Note, not all people use full-sized windows all the time, so the
widths here must not only be considered in terms of display
resolutions...

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>