ui: node/status view: reduce quite generous paddings
As both, the whole panels body and the actual items all had their own
padding it added up quite a bit.

Some padding is good to avoid elements being "glued" to the parent
borders, but that can be done with also with ~15 px vs. 30 px on each
side, so no need to waste that much extra space we can use in some
languages to render content

Added benefit, the status panel now is more in sync with the RRD
panels regarding content start/end.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>