fix #3597: ui: ceph setup: avoid autofill of monitor network
By not auto filling the Ceph public network we can avoid accidental
clicks on 'Next' which will cause the first Mon to be created with a
potentially wrong network. While that is fixable, it is tedious and
can be easily avoided by making the user always select the network to
use.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Tested-by: Stefan Hrdlicka<s.hrdlicka@proxmox.com>
 [ T: adapted commit subject to be more specific and match our common
   style ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>