ui: node: rework layout of certificate viewer
Move the raw cert into a fieldset and collapse that by default, also
make it render monospace and pre-wrap whitespace.

Use a inputpanel to get top, col 1/2 & bottom layout so that rsa key
type/size and validity range can be rendered more compactly

n addition to that limit the whole windows height but make it
scrollable, this allows to sanely use the viewer on certs with many
SANs even on small window/browser sizes.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>