pveceph: add osd details command
To provide similar output on the CLI as is possible in the GUI/API
regaring OSD details.

By default (output-format=text) a more concise output is shown. Using
json or yaml as output format will print all the available data.

The 'verbose' flag causes json-pretty output to be used.

The functionality is split between the actual function and the output
formatter as not all options/parameters are available in each.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>