fix #4194: ui: clarify that it's actually a bulk shutdown, not a stop
As some users where confused by the usage of "Stop", which we
normally reserve for a hard-stop.

And yes, while the bulk shutdown formerly always had a timeout armed
hard-stop hard coded, it was still subtle and we recently exposed
control on that via the API an UI, so use shutdown to be more in line
with the CT/VM naming, e.g., in their power menus.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>