pve7to8: Improve systemd unified cgroupv2 support check
Checking /lib/systemd if it is present and a directory is not enough, as
the shared object file used to check the version might nevertheless be
located at /usr/lib/systemd, or under /usr/lib/x86_64-linux-gnu/systemd.
So check also the latter paths, if the former returned no match.

Further, Arch Linux appends the minor version and release version to the
filename, so include that in the regex as well.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>