pve7to8: make check for correct running kernel more resilient
users can have the 6.2 kernel already installed, so this check needs
to cope with that.

The last non-bpo version on bullseye releases was 6.2.11-2-pve and
the first release for bookworm was 6.2.16-1-pve, so use a regex that
matches the bookworm one and newer, but fails with any tilde
characters, which must be used for any 6.2 update in bullseye from
now on to ensure correct upgrades.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>