pve7to8: nvidia vGPU: downgrad report for active service to warning
While it would be a failure now (and I thus have recommended Dominik
to use a fail here), we cannot really predict how this
evolves. If NVIDIA fixes the issue in newer driver/tooling we cannot
detect this and telling users to ignore a failure can lead to them
getting condition on that and start ignoring any failure.

So, let's ignore my initial error-level hunch and use a warning,
which is quite prominent too.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>